doc_backend_project/requirements.txt
# ===== Core backend =====
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.0
requests==2.31.0

# ===== Document handling =====
PyMuPDF==1.24.8
python-docx==1.1.0
python-pptx==0.6.23

# ===== Tables =====
camelot-py==0.11.0
tabula-py==2.8.0

# ===== OCR =====
pytesseract==0.3.13
Pillow==10.4.0

# ===== Image captioning / Transformers (optional) =====
transformers==4.40.2

# ===== Embeddings & vector DB =====
sentence-transformers==2.2.2
faiss-cpu==1.11.0

# ===== Utilities =====
opencv-python==4.12.0.88

# ===== Dev / helpers (optional) =====
python-multipart==0.0.6   # if you use form file uploads in FastAPI
python-dotenv==1.0.0
